faad: fixes warning in macosx snow leopard

+++ ext/faad/gstfaad.c
@@ -346,7 +346,7 @@ gst_faad_setcaps (GstPad * pad, GstCaps * caps)
     }
 
     GST_DEBUG_OBJECT (faad, "codec_data init: channels=%u, rate=%u", channels,
-        samplerate);
+        (guint32) samplerate);
 
     /* not updating these here, so they are updated in the
      * chain function, and new caps are created etc. */
@@ -1310,15 +1310,15 @@ gst_faad_chain (GstPad * pad, GstBuffer * buffer)
       if (faacDecInit (faad->handle, input_data, input_size, &rate, &ch) < 0)
         goto init_failed;
 
-      GST_DEBUG_OBJECT (faad, "faacDecInit() ok: rate=%u,channels=%u", rate,
-          ch);
+      GST_DEBUG_OBJECT (faad, "faacDecInit() ok: rate=%u,channels=%u",
+          (guint32) rate, ch);
     } else {
       if ((gint8) faacDecInit2 (faad->handle, faad->fake_codec_data, 2,
               &rate, &ch) < 0) {
         goto init2_failed;
       }
-      GST_DEBUG_OBJECT (faad, "faacDecInit2() ok: rate=%u,channels=%u", rate,
-          ch);
+      GST_DEBUG_OBJECT (faad, "faacDecInit2() ok: rate=%u,channels=%u",
+          (guint32) rate, ch);
     }
 
     skip_bytes = 0;
